refactor trace sampling into specific package
