+++ dynamic-control/src/test/java/io/opentelemetry/contrib/dynamic/policy/tracesampling/DelegatingSamplerTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.contrib.dynamic.sampler;
+package io.opentelemetry.contrib.dynamic.policy.tracesampling;
 
 import static org.assertj.core.api.Assertions.assertThat;
